requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
numpy==2.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
